test: add unit tests for following tab
