+++ test/nuxt/pages/home/index.spec.ts
@@ -1,177 +1,30 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { mount } from '@vue/test-utils';
-import { nextTick } from 'vue';
+import { mockNuxtImport, mountSuspended } from '@nuxt/test-utils/runtime';
 
-const { apiFetchMock } = vi.hoisted(() => ({
-  apiFetchMock: vi.fn(),
-}));
-
-vi.mock('~/api', () => ({
-  apiFetch: apiFetchMock,
-}));
-
-vi.stubGlobal('definePageMeta', () => {});
-vi.stubGlobal('$t', (k: string) => k);
-
-vi.mock('@vueuse/core', () => {
+const { navigateToMock } = vi.hoisted(() => {
   return {
-    useInfiniteScroll: () => {},
-    useVirtualList: (list: unknown) => ({ list, containerProps: {}, wrapperProps: {} }),
+    navigateToMock: vi.fn(),
   };
 });
 
-describe('Home page (unit)', () => {
+mockNuxtImport('navigateTo', () => navigateToMock);
+
+describe('Home Index Page', () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
 
-  it('loads tweets and renders TweetDefaultCard items', async () => {
-    const mockTweet = {
-      id: 'tw-1',
-      author: { username: 'u1', displayName: 'U One', avatarUrl: '' },
-      content: 'hello',
-      createdAt: new Date().toISOString(),
-      replyCount: 0,
-      retweetCount: 0,
-      likeCount: 0,
-      isLiked: false,
-      isRetweeted: false,
-      entities: { mentions: [], hashtags: [] },
-      media: [],
-    };
+  it('redirects to /home/for-you with code 301', async () => {
+    const { default: HomePage } = await import('~/pages/home/index.vue');
+    await mountSuspended(HomePage);
 
-    apiFetchMock.mockResolvedValue({
-      data: {
-        data: [mockTweet],
-        pagination: { cursor: '0', nextCursor: null, hasNextPage: false },
-      },
-    });
-
-    const HomePage = (await import('@/pages/home/index.vue')).default;
-
-    const wrapper = mount(HomePage, {
-      global: {
-        stubs: { TweetDefaultCard: true },
-        mocks: { $t: (k: string) => k },
-      },
-    });
-
-    await nextTick();
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(apiFetchMock).toHaveBeenCalled();
-
-    const cards = wrapper.findAllComponents({ name: 'TweetDefaultCard' });
-    expect(cards.length).toBeGreaterThanOrEqual(1);
+    expect(navigateToMock).toHaveBeenCalledWith('/home/for-you', { redirectCode: 301 });
   });
 
-  it('handles empty response (no new tweets) without rendering items', async () => {
-    apiFetchMock.mockResolvedValue({
-      data: { data: [], pagination: { cursor: null, nextCursor: null, hasNextPage: false } },
-    });
+  it('calls navigateTo exactly once', async () => {
+    const { default: HomePage } = await import('~/pages/home/index.vue');
+    await mountSuspended(HomePage);
 
-    const HomePage = (await import('@/pages/home/index.vue')).default;
-
-    const wrapper = mount(HomePage, {
-      global: {
-        stubs: { TweetDefaultCard: true },
-        mocks: { $t: (k: string) => k },
-      },
-    });
-
-    await nextTick();
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(apiFetchMock).toHaveBeenCalled();
-
-    const cards = wrapper.findAllComponents({ name: 'TweetDefaultCard' });
-    expect(cards.length).toBe(0);
-  });
-
-  it('logs error and resets loading when fetch fails', async () => {
-    const error = new Error('network');
-    apiFetchMock.mockRejectedValue(error);
-
-    const consoleSpy = vi.spyOn(console, 'error').mockImplementation(() => {});
-
-    const HomePage = (await import('@/pages/home/index.vue')).default;
-
-    const wrapper = mount(HomePage, {
-      global: {
-        stubs: { TweetDefaultCard: true },
-        mocks: { $t: (k: string) => k },
-      },
-    });
-
-    await nextTick();
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(apiFetchMock).toHaveBeenCalled();
-    expect(consoleSpy).toHaveBeenCalled();
-
-    expect(wrapper.find('.text-muted-foreground').exists()).toBe(false);
-
-    consoleSpy.mockRestore();
-  });
-
-  it('invokes the infinite-scroll handler provided to useInfiniteScroll', async () => {
-    vi.doMock('@vueuse/core', () => {
-      return {
-        useInfiniteScroll: (_el: unknown, cb: () => unknown) => {
-          void cb();
-        },
-        useVirtualList: (list: unknown) => ({ list, containerProps: {}, wrapperProps: {} }),
-      };
-    });
-
-    apiFetchMock.mockResolvedValue({
-      data: { data: [], pagination: { cursor: null, nextCursor: null, hasNextPage: false } },
-    });
-
-    const HomePage = (await import('@/pages/home/index.vue')).default;
-
-    mount(HomePage, {
-      global: {
-        stubs: { TweetDefaultCard: true },
-        mocks: { $t: (k: string) => k },
-      },
-    });
-
-    await nextTick();
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(apiFetchMock).toHaveBeenCalled();
-  });
-
-  it('executes canLoadMore option from useInfiniteScroll to cover reactive check', async () => {
-    vi.doMock('@vueuse/core', () => {
-      return {
-        useInfiniteScroll: (
-          _el: unknown,
-          cb: () => unknown,
-          options: { canLoadMore?: () => boolean } | undefined,
-        ) => {
-          // Invoke canLoadMore to evaluate `hasNextPage.value && !isLoading.value`
-          if (options?.canLoadMore) options.canLoadMore();
-          void cb();
-        },
-        useVirtualList: (list: unknown) => ({ list, containerProps: {}, wrapperProps: {} }),
-      };
-    });
-
-    apiFetchMock.mockResolvedValue({
-      data: { data: [], pagination: { cursor: null, nextCursor: null, hasNextPage: false } },
-    });
-
-    const HomePage = (await import('@/pages/home/index.vue')).default;
-
-    mount(HomePage, {
-      global: { stubs: { TweetDefaultCard: true }, mocks: { $t: (k: string) => k } },
-    });
-
-    await nextTick();
-    await new Promise((r) => setTimeout(r, 0));
-
-    expect(apiFetchMock).toHaveBeenCalled();
+    expect(navigateToMock).toHaveBeenCalledTimes(1);
   });
 });
